Move `Txid`
