+++ stacks-codec/Cargo.toml
@@ -16,5 +16,6 @@ edition = "2021"
 
 [dependencies]
 stacks-common = { path = "../stacks-common", default-features = false }
+clarity-types = { path = "../clarity-types", default-features = false }
 serde = { workspace = true, features = ["derive"] }
