+++ stackslib/src/chainstate/stacks/mod.rs
@@ -32,7 +32,6 @@ use serde_json::json;
 #[allow(unused_imports)]
 use stacks_common::address::AddressHashMode;
 use stacks_common::codec::Error as codec_error;
-use stacks_common::deps_common::bitcoin::util::hash::Sha256dHash;
 #[cfg(test)]
 use stacks_common::types::chainstate::StacksBlockId;
 use stacks_common::types::chainstate::{
@@ -382,38 +381,6 @@ impl Error {
                 Error::CostOverflowError(cost_before, cur_cost, budget)
             }
         }
-    }
-}
-
-impl Txid {
-    /// A Stacks transaction ID is a sha512/256 hash (not a double-sha256 hash)
-    pub fn from_stacks_tx(txdata: &[u8]) -> Txid {
-        let h = Sha512Trunc256Sum::from_data(txdata);
-        let mut bytes = [0u8; 32];
-        bytes.copy_from_slice(h.as_bytes());
-        Txid(bytes)
-    }
-
-    /// A sighash is calculated the same way as a txid
-    pub fn from_sighash_bytes(txdata: &[u8]) -> Txid {
-        Txid::from_stacks_tx(txdata)
-    }
-
-    /// Create a [`Txid`] from the tx hash bytes used in bitcoin.
-    /// This just reverses the inner bytes of the input.
-    pub fn from_bitcoin_tx_hash(tx_hash: &Sha256dHash) -> Txid {
-        let mut txid_bytes = tx_hash.0;
-        txid_bytes.reverse();
-        Self(txid_bytes)
-    }
-
-    /// Create a [`Sha256dHash`] from a [`Txid`]
-    /// This assumes the inner bytes are stored in "big-endian" (following the hex bitcoin string),
-    /// so just reverse them to properly create a tx hash.
-    pub fn to_bitcoin_tx_hash(txid: &Txid) -> Sha256dHash {
-        let mut txid_bytes = txid.0;
-        txid_bytes.reverse();
-        Sha256dHash(txid_bytes)
     }
 }
